JSON parsing brought out into the separate class. Slight modifications to the left menu styles done.
