+++ app/src/main/java/ru/homecatering/LeftMenuAdapter.java
@@ -2,9 +2,6 @@ package ru.homecatering;
 
 import android.app.Activity;
 import android.content.Context;
-import android.graphics.drawable.Drawable;
-import android.util.JsonReader;
-import android.util.JsonToken;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -13,92 +10,17 @@ import android.widget.BaseExpandableListAdapter;
 import android.widget.ImageView;
 import android.widget.TextView;
 
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.nio.charset.StandardCharsets;
-import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
 class LeftMenuAdapter extends BaseExpandableListAdapter {
-    private Activity activity;
-    private List<LeftMenuModel> lmHeader = new ArrayList<>();
-    private Map<LeftMenuModel, List<LeftMenuModel>> lmChildren = new HashMap<>();
+    private List<LeftMenuModel> lmHeader;
+    private Map<LeftMenuModel, List<LeftMenuModel>> lmChildren;
 
     LeftMenuAdapter(Activity activity) {
-        this.activity = activity;
-        try {
-            InputStream is = activity.getResources().openRawResource(R.raw.left_menu);
-            JsonReader reader = new JsonReader(new InputStreamReader(is, StandardCharsets.UTF_8));
-            readOuter(reader);
-            is.close();
-        } catch (IOException e) {
-            Log.e("JSON_ERROR", "Error while reading json.", e);
-        }
-    }
-
-    private void readOuter(JsonReader reader) throws IOException {
-        reader.beginArray();
-        while (reader.hasNext()) {
-            String id = "", text = "";
-            Drawable drawable = null;
-            List<LeftMenuModel> lmc = null;
-            reader.beginObject();
-            while (reader.hasNext()) {
-                switch (reader.nextName()) {
-                    case "id":
-                        id = reader.nextString();
-                        break;
-                    case "text":
-                        text = activity.getString(activity.getResources().getIdentifier(reader.nextString(), "string", activity.getPackageName()));
-                        break;
-                    case "icon":
-                        drawable = activity.getDrawable(activity.getResources().getIdentifier(reader.nextString(), "drawable", activity.getPackageName()));
-                        break;
-                    case "children":
-                        if (reader.peek() != JsonToken.NULL) {
-                            lmc = readChild(reader);
-                        } else {
-                            reader.skipValue();
-                        }
-                        break;
-                }
-            }
-            reader.endObject();
-            LeftMenuModel lmOuter = new LeftMenuModel(text, id, drawable, lmc != null);
-            lmHeader.add(lmOuter);
-            lmChildren.put(lmOuter, lmc);
-        }
-        reader.endArray();
-    }
-
-    private List<LeftMenuModel> readChild(JsonReader reader) throws IOException {
-        List<LeftMenuModel> lmc = new ArrayList<>();
-        String id = "", text = "";
-        Drawable drawable = null;
-        reader.beginArray();
-        while (reader.hasNext()) {
-            reader.beginObject();
-            while (reader.hasNext()) {
-                switch (reader.nextName()) {
-                    case "id":
-                        id = reader.nextString();
-                        break;
-                    case "text":
-                        text = activity.getString(activity.getResources().getIdentifier(reader.nextString(), "string", activity.getPackageName()));
-                        break;
-                    case "icon":
-                        drawable = activity.getDrawable(activity.getResources().getIdentifier(reader.nextString(), "drawable", activity.getPackageName()));
-                        break;
-                }
-            }
-            reader.endObject();
-            lmc.add(new LeftMenuModel(text, id, drawable, false));
-        }
-        reader.endArray();
-        return lmc;
+        LeftMenuJSONParser leftMenuData = new LeftMenuJSONParser(activity);
+        lmHeader = leftMenuData.getMenuHeader();
+        lmChildren = leftMenuData.getMenuChildren();
     }
 
     @Override
@@ -141,7 +63,7 @@ class LeftMenuAdapter extends BaseExpandableListAdapter {
     public View getGroupView(int groupPosition, boolean isExpanded, View convertView, ViewGroup parent) {
         LeftMenuModel lmm = getGroup(groupPosition);
         if (convertView == null) {
-            LayoutInflater inflater = (LayoutInflater) activity.getSystemService(Context.LAYOUT_INFLATER_SERVICE);
+            LayoutInflater inflater = (LayoutInflater) parent.getContext().getSystemService(Context.LAYOUT_INFLATER_SERVICE);
             convertView = inflater.inflate(R.layout.left_menu_item, null);
         }
 
@@ -152,6 +74,7 @@ class LeftMenuAdapter extends BaseExpandableListAdapter {
         TextView text = convertView.findViewById(R.id.text);
         text.setText(lmm.getText());
         if (lmm.hasChildren) {
+            Log.i("INFO",lmm.getText()+lmm.hasChildren+isExpanded);
             if (isExpanded) {
                 convertView.findViewById(R.id.more).setVisibility(View.INVISIBLE);
                 convertView.findViewById(R.id.less).setVisibility(View.VISIBLE);
@@ -167,7 +90,7 @@ class LeftMenuAdapter extends BaseExpandableListAdapter {
     public View getChildView(int groupPosition, int childPosition, boolean isLastChild, View convertView, ViewGroup parent) {
         LeftMenuModel lmm = getChild(groupPosition, childPosition);
         if (convertView == null) {
-            LayoutInflater inflater = (LayoutInflater) activity.getSystemService(Context.LAYOUT_INFLATER_SERVICE);
+            LayoutInflater inflater = (LayoutInflater) parent.getContext().getSystemService(Context.LAYOUT_INFLATER_SERVICE);
             convertView = inflater.inflate(R.layout.left_menu_item_inner, null);
         }
